Added Authorization: Bearer header to the request
I have faced a problem where LinkedIn would return `Unknown authentication scheme` because the token was only included in the HTTP parameters. However, the official documentation states that to successfully sign requests, we have to include an Authorization header:

> Once you've obtained an Access Token, you can start making authenticated API requests on behalf of the user. This is accomplished by including an "Authorization" header in your HTTP call to LinkedIn's API.

So I edited the class to include the header in the fetch method.